Move MoreInfoModal to a new component
