+++ assets/src/scripts/components/TreeRow.jsx
@@ -3,58 +3,9 @@ import React, { useState } from "react";
 import { Button, ButtonGroup, Modal } from "react-bootstrap";
 import { readValueFromPage } from "../_utils";
 import DescendantToggle from "./DescendantToggle";
+import MoreInfoModal from "./MoreInfoModal";
 import Pipes from "./Pipes";
 import StatusToggle from "./StatusToggle";
-
-function createModalText({
-  allCodes,
-  code,
-  codeToStatus,
-  codeToTerm,
-  hierarchy,
-  status,
-}) {
-  const included = allCodes.filter((c) => codeToStatus[c] === "+");
-  const excluded = allCodes.filter((c) => codeToStatus[c] === "-");
-  const significantAncestors = hierarchy.significantAncestors(
-    code,
-    included,
-    excluded,
-  );
-
-  const includedAncestorsText = significantAncestors.includedAncestors
-    .map((code) => `${codeToTerm[code]} (${code})`)
-    .join(", ");
-
-  const excludedAncestorsText = significantAncestors.excludedAncestors
-    .map((code) => `${codeToTerm[code]} (${code})`)
-    .join(", ");
-
-  let text = "";
-
-  switch (status) {
-    case "+":
-      text = "Included";
-      break;
-    case "(+)":
-      text = `Included by ${includedAncestorsText}`;
-      break;
-    case "-":
-      text = "Excluded";
-      break;
-    case "(-)":
-      text = `Excluded by ${excludedAncestorsText}`;
-      break;
-    case "?":
-      text = "Unresolved";
-      break;
-    case "!":
-      text = `In conflict!  Included by ${includedAncestorsText}, and excluded by ${excludedAncestorsText}`;
-      break;
-  }
-
-  return text;
-}
 
 function TreeRow({
   allCodes,
@@ -82,18 +33,6 @@ function TreeRow({
 
   const rowSpacing = pipes.length === 0 ? "mt-2" : "mt-0";
   const className = `${rowSpacing} d-flex`;
-
-  const modalText = isEditable
-    ? createModalText({
-        allCodes,
-        code,
-        codeToStatus,
-        codeToTerm,
-        hierarchy,
-        status,
-      })
-    : "";
-  const [showMoreInfoModal, setShowMoreInfoModal] = useState(false);
 
   return (
     <div className={className} data-code={code} data-path={path}>
@@ -128,27 +67,15 @@ function TreeRow({
       </div>
 
       {isEditable ? (
-        <>
-          <Button
-            className="py-0 border-0"
-            onClick={() => setShowMoreInfoModal(true)}
-            variant="outline-secondary"
-          >
-            &hellip;
-          </Button>
-
-          <Modal
-            show={showMoreInfoModal}
-            onHide={() => setShowMoreInfoModal(false)}
-          >
-            <Modal.Header closeButton>
-              <Modal.Title>
-                {term} ({code})
-              </Modal.Title>
-            </Modal.Header>
-            <Modal.Body>{modalText}</Modal.Body>
-          </Modal>
-        </>
+        <MoreInfoModal
+          allCodes={allCodes}
+          code={code}
+          codeToStatus={codeToStatus}
+          codeToTerm={codeToTerm}
+          hierarchy={hierarchy}
+          status={status}
+          term={term}
+        />
       ) : null}
     </div>
   );
